requirements.txt
# 基础依赖
aiohttp>=3.8.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
PyYAML>=6.0
python-dotenv>=0.19.0

# 数据库相关
redis>=4.3.0
motor>=3.1.1
pymongo>=4.3.3
faiss-cpu>=1.7.0

# AI/ML相关
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# 工具和实用程序
python-jose>=3.3.0
python-multipart>=0.0.5
jieba>=0.42.1
numpy>=1.21.0
loguru>=0.6.0

# 开发工具
black>=22.3.0
isort>=5.10.1
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
aiosqlite>=0.17.0
Flask>=2.0.1
